requirements.txt
python-telegram-bot>=20.0
requests>=2.28.0
python-dotenv>=0.21.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.3.0
colorama>=0.4.0
tqdm>=4.65.0
python-dateutil>=2.8.2
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0
gunicorn>=20.1.0
uvicorn>=0.20.0
